requirements.txt
alembic
fastapi
psycopg2
python-decouple
pytest
requests
sqlmodel
uvicorn